% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{centroids_tidy_format}
\alias{centroids_tidy_format}
\title{Convert centroids matrix into tidy format}
\usage{
centroids_tidy_format(x)
}
\arguments{
\item{x}{an object returned by ClustImpute}
}
\value{
data frame with three columns ( cluster | Features | value ) that uniquely defines the centroids#'
}
\description{
A helper function for plot() but may be useful for further post-processing of clustering results.
}
