% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{check_replace_dups}
\alias{check_replace_dups}
\title{Check and replace duplicate (centroid) rows}
\usage{
check_replace_dups(centroids, X, seed)
}
\arguments{
\item{centroids}{Matrix of centroids}

\item{X}{Underlying data matrix (without missings)}

\item{seed}{Seed used for random sampling}
}
\value{
Returns centroids where duplicate rows are replaced by random draws
}
\description{
Internal function of ClustImpute: check new centroids for duplicate rows and replace with random draws in this case.
}
