% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{plot.kmeans_ClustImpute}
\alias{plot.kmeans_ClustImpute}
\title{Plot showing marginal distribution by cluster assignment}
\usage{
\method{plot}{kmeans_ClustImpute}(
  x,
  type = "hist",
  vline = "centroids",
  hist_bins = 30,
  color_bins = "#56B4E9",
  color_vline = "#E69F00",
  size_vline = 2,
  ...
)
}
\arguments{
\item{x}{an object returned from ClustImpute}

\item{type}{either "hist" to plot a histogram or "box" for a boxplot}

\item{vline}{for "hist" a vertical line is plotted showing either the centroid value or the mean of all data points grouped by cluster and feature}

\item{hist_bins}{number of bins for histogram}

\item{color_bins}{color for the histogram bins}

\item{color_vline}{color for the vertical line}

\item{size_vline}{size of the vertical line}

\item{...}{currently unused}
}
\value{
Returns a ggplot object
}
\description{
Returns a plot with the marginal distributions by cluster and feature. The plot is a ggplot object an can therefore be further modified.
}
