% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{predict.kmeans_ClustImpute}
\alias{predict.kmeans_ClustImpute}
\title{Prediction method}
\usage{
\method{predict}{kmeans_ClustImpute}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class kmeans_ClustImpute}

\item{newdata}{Data frame}

\item{...}{additional arguments affecting the predictions produced - not currently used}
}
\value{
integer value (cluster assignment)
}
\description{
Prediction method
}
