% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{print.kmeans_ClustImpute}
\alias{print.kmeans_ClustImpute}
\title{Print method for ClustImpute}
\usage{
\method{print}{kmeans_ClustImpute}(x, ...)
}
\arguments{
\item{x}{an object returned from ClustImpute}

\item{...}{currently unused}
}
\description{
Returns a plot with the marginal distributions by cluster and feature. The plot shows histograms or boxplots and , as a ggplot object, can be modified further.
}
