% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{ClustImpute}
\alias{ClustImpute}
\title{K-means clustering with build-in missing data imputation}
\usage{
ClustImpute(X, nr_cluster, nr_iter = 10, c_steps = 1,
  wf = default_wf, n_end = 10, seed_nr = 150519)
}
\arguments{
\item{X}{Data frame with only numeric values or NAs}

\item{nr_cluster}{Number of clusters}

\item{nr_iter}{Iterations of procedure}

\item{c_steps}{Number of clustering steps per iteration}

\item{wf}{Weight function. linear up to n_end by default}

\item{n_end}{Steps until convergence of weight function to 1}

\item{seed_nr}{Number for set.seed()}
}
\value{
\describe{
  \item{complete_data}{Completed data without NAs}
  \item{clusters}{For each row of complete_data, the associated cluster}
  \item{centroids}{For each cluster, the coordinates of the centroids}
  \item{imp_values_mean}{Mean of the imputed variables per draw}
  \item{imp_values_sd}{Standard deviation of the imputed variables per draw}
}
}
\description{
Clustering algorithm that produces a missing value imputation using  on the go. The (local) imputation distribution is defined by the currently assigned cluster. The first draw is by random imputation.
}
