% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{default_wf}
\alias{default_wf}
\title{K-means clustering with build-in missing data imputation}
\usage{
default_wf(n, n_end = 10)
}
\arguments{
\item{n}{current step}

\item{n_end}{steps until convergence of weight function to 1}
}
\value{
value between 0 and 1
}
\description{
Default weight function
}
