% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustImpute.R
\name{var_reduction}
\alias{var_reduction}
\title{Reduction of variance}
\usage{
var_reduction(clusterObj)
}
\arguments{
\item{clusterObj}{Object of class kmeans_ClustImpute}
}
\value{
integer value typically between 0 and 1
}
\description{
Computes one minus the ratio of the sum of all within cluster variances by the overall variance
}
